+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Unexpected closing when disconnecting from an EtherCAT (SDOs) drive if servo status listener is active.
 - Avoid crashes in the Poller due to read timeouts.
 - Poller timer thread not closing after the poller is finished.
+- Improve the enable and disable methods of the Servo class.
 
 ### Changed
 - Raise ILValueError when the disturbance data does not fit in a register.